+++ R/deseq2pip-package.R
@@ -58,5 +58,6 @@
 #' }
 #'
 #' @importFrom SummarizedExperiment colData `colData<-`
+#' @importFrom S4Vectors DataFrame
 #' @name deseq2pip-package
 NULL